CI: Fix static checks script for callers
The static checks script was incorrectly assuming it was being run in
the tests repository. This "works", but then only runs the checks on the
`tests` repository, *not* those of the caller (which might actually be the
`tests` repository of course). Crucially, callers need to be able to
invoke the static check script without `cd`-ing to the `tests`
repository directory.

Fixes #13.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>