CI: Support KATA_DEV_MODE to enable lenient linters
The static checks script currently only runs a subset of linters under
the CI, but all linters by default otherwise.

Added a `KATA_DEV_MODE` variable that, if set, will run the same subset.
This is useful and crucially does not require the developer to pretend
to be running under a CI (which could be dangerous).

Fixes #24.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>