terminal: Properly set raw terminal only when needed
This commit relies on unix package from golang instead of moby/moby
to get the terminal setup in raw mode.

Also, it fixes an issue about the terminal being set in raw mode
every time, even when STDIN was not a terminal. A new "terminal"
flag has been introduced to specify if a terminal has been setup.
Based on this information, the shim will try (or not) to set this
terminal to raw mode.

Fixes #22

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>