log: Discard logs going to stderr
By default, the logrus logger redirect the logs to stderr of the
shim. This happens even when the syslog hook is added, meaning logs
are both printed to stderr and the system journal.

We cannot let this happen since those logs should not be routed back
to the output of the shim, giving some weird output while you only
expect the output coming from the command executed on the container.

That's why this commit redirects those logs to a fake io.Writer called
devNull, defining a null operation, and returning instantly when the
writer is called. This leaves only the system journal being written.

Fixes #46

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>