CI: Fix handling of sub-modules in static checks script
Don't ignore a repository just because it contains sub-modules - filter
the packages those sub-modules provide out of the list of all packages
and run the static checks on the remaining packages.

This is specifically for the https://github.com/kata-containers/runtime
repo. It used to contain simply submodule references to other
repositories. However, now it contains actual code content (`cli/` and
`virtcontainers/`) but the original check script logic was ignoring
*all* packages below runtime. This is incorrect since (now) both the
`cli` and `virtcontainers` packages should be checked when a runtime PR
is raised.

Fixes #136.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>