metrics: checkmetrics: add checkmetrics cmd
The `checkmetrics` cmd is used to process the metrics
JSON results files and compare them against a set of
`baseline` metrics, generate a report summary, and
return a failure code depending on if all the tests hit
their metrics targets.

The command currently supports both JSON and the older
CSV results. CSV results processing is deprecated, and will
be phased out.

Signed-off-by: Graham whaley <graham.whaley@intel.com>