ci: crio: only install go-md2man if it does not exist
If we are re-running on a dirty tree then we may have already
installed go-md2man, in which case the link fails as we then
try to link the same file to itself.
If the program is already installed in the GOBIN, do not try
to install it again.

With this last commit of the series:
Fixes: #216

Signed-off-by: Graham whaley <graham.whaley@intel.com>