CI: add support for new cri-o job
This change is needed because now that we plan to test
the changes in the cri-o repository, the CI needs to know
if the change being tested comes from the cri-o repository
or from our kata-containers repos. If it comes from the cri-o
repository, we should not use the version we support on our
versions.yaml. We should stick to the changes in the PR.

Fixes: #230.

Signed-off-by: Salvador Fuentes <salvador.fuentes@intel.com>