metrics: launch_times: set runtime for background containers
We were not using the defined RUNTIME when launching the background
(scaling test) containers. This meant the scaling test only really
measured what we thought when the runtime we were testing was the
docker default one.

Explicitly state which runtime to use for the background containers
so we test what we expect for any setting of RUNTIME.

Fixes: #239

Signed-off-by: Graham whaley <graham.whaley@intel.com>