log-parser: ignore badly formed agent messages by default
Both the agent and the kernel write to the console. The log parser
expects all writes to the console *by the agent* to be structured and
well-formed. However, if the kernel writes to the console at the same
time the agent is writing to it, the resulting log entry will appear to
the log parser as "corrupt".

Make the log parser warn if such issues are found rather than failing.
For the original behaviour, run the log parser with the new `--strict`
option. To hide warnings, specify the new `--quiet` option.

Fixes #228.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>