CI: Run k8s tests on ubuntu and Openshift tests on Fedora
As we need CRI-O 1.9 to run tests on openshift v3.9.0 (latest)
and CRI-O 1.10 to run Kubernetes 1.10, we need to make a
division in our CI, so that we can test both scenarios without
adding more execution time to the CI jobs.
Because of this, we will execute CRI-O 1.9 - openshift v3.9.0
tests under Fedora and CRI-O 1.10 - Kubernetes 1.10 tests
under ubuntu.

Fixes: #276.

Depends-on: github.com/kata-containers/runtime#278

Signed-off-by: Salvador Fuentes <salvador.fuentes@intel.com>