CRI-O: Use block device for devicemapper driver
Now that cri-o is deprecating the use of loop back
devices when using devicemapper driver, we need to
use a block device on the Azure VMs from our CI.

This fix assumes that the Azure VM will have an
/dev/sdb device and it will only be executed when
CI=true and when the KATA_DEV_MODE is not set.
In addition it also checks that the block device
does not have any partitions before running the
tests.

Fixes: #341

Signed-off-by: Salvador Fuentes <salvador.fuentes@intel.com>