scripts: Add arch script
The standard `arch(1)` utility is rather limited in that it only shows
the host architecture in a single format.

Various repos need to know the architecture in either "golang" format
or using the name the Linux kernel config system uses.

Created a new script that provides this information for the currently
supported architectures that can be used by repos without having to
hard-code the logic themselves (or download `golang`).

Fixes #382.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>