kata-manager: Add no execute option
Added a `-n` option for `kata-manager.sh` to generate the script(s)
only (don't run them). This allows users to review before executing
them.

This necessitated modifying `kata-doc-to-script.sh` to allow a comment
to be added into the script since `kata-manager.sh` can generate
multiple scripts and it is useful for users to know which script does
what.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>