ci: install packaged image
This commit will install the packaged image
if it is up to date. And will also try to update
the agent by mounting the image in a local path.
If the process cannot be completed successfully,
the script will build the image from sources.

This approach should reduce the setup time spent
for setting up the environment.

Fixes #410.

Signed-off-by: Salvador Fuentes <salvador.fuentes@intel.com>