CI: Only check URL changed by PRs
Change the URL checker to only check the URLs inside document files modified
by a PR. This is relatively difficult since if a PR adds a new document and
modifies an existing document that references that new document, the URL for
the new document will not be valid. The checker handles this by constructing
the correct URL for this new document which will be valid once the PR lands.
As long as the new-but-not-yet-valid URL referenced by the modified document
matches what the checker thinks the URL will be, the URL is accepted.

For the master branch, all doc URLs are checked.

This change necessitated modifying how the static check script is called by
the CI.

Fixes #409.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>