proxy: Prevent from socket connection leak
The proxy was cleaning up the connection with both sides (that is
with shim/runtime from one side, and VM/agent from the other side)
only if both connections were closed from outside. When one of the
connection was still up (which was not making a lot of sense since
it was not connected to the other end), proxy go routine was still
waiting and the connection was never ended up.

In order to prevent from this kind of behavior, the proxy will close
connections previously opened as soon as one end of the connection
returns. This way, it won't leak socket connection remaining opened.

Fixes #77

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>