checkmetrics: Improve calculation of 'gap %'
Previously we presented the gap % as the % overhead based
on the minimum value. Now we are heading towards enabling
an actual % based output option, it makes more sense to
show the % gap as the % spread out from the 'mid point'
of the acceptible range.
Also modify tests to take into account the change.

Signed-off-by: Graham Whaley <graham.whaley@intel.com>