CI: Add procenv output to teardown
Run the `procenv(1)` tool (which dumps as much information about it's
environment as possible) twice at the end of the CI: once as the CI
user and once as `root`. This could be useful to diagnose test failures
by comparing its CI output with the output generated on a local system.

Fixes #556.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>