metrics: use docker kill rather than docker stop
For many of our 'used' test containers, issuing a `docker stop`
actually results in the stop timing out (as it cannot signal the
container to stop for any one of a number of potential reasons),
and then docker reverts to a kill anyhow. The cost of this is a
10s delay per `stop`, whilst docker times out and then issues the
kill. That costs us heavily when we are doing multi-container tests
using say 20 or 100 containers, as each one takes a 10s overhead
on stop/kill/removing it.

Move all the metrics test `stops` to be `kills`.

Fixes: #561

Signed-off-by: Graham Whaley <graham.whaley@intel.com>