metrics: reportgen: Add metrics report generation scripts
Add a set of scripts and Dockerfiles that allow metrics data to
be gathered, processed and presented in a mostly automated and
repeatable manner.

The package provides:
 - A script to gather metrics data for the system under test
 - A set of Dockerfile/RMarkdown/R files for processing and presenting
  the gathered data
 - A script to drive the Dockerfile/R files to generate a PDF report.

The scripts can process 1-n sets of data. If n=2 then some parts of
the report will add an additional set of 'diff' data for easier
comparison.

Fixes: #596

Signed-off-by: Graham Whaley <graham.whaley@intel.com>