metrics: report: make grabdata modular
Allow configuration, via the commandline, of which tests
the grabdata.sh script runs. By default it will run them
all.

Added as some of the tests can take a long time, and sometimes
you are really only interested in one or a few of the tests. To
save having to edit the test before running, add a test selection
capability to the commandline.

Signed-off-by: Graham Whaley <graham.whaley@intel.com>