metrics: ci: enable cloud instance metrics CI
If we are on a known noisy/time variant system then we can still
run some metrics that are nominally time insensitive (like the
footprint/density tests). Modify the scripts so we can set an
env var to notify we are on such a system, and thus adapt the
tests we run, and how we locate a correct checkmetrics file to
use to check the results.

Fixes: #698

Signed-off-by: Graham Whaley <graham.whaley@intel.com>