metrics: report: fix subdir check in makereport
An early change slide in a stderr re-direct that broke the check
in the good case as well.
Fix that by moving the stderr redirect. Also, drop the chdir as it
is not needed, and can fail in itself.
Whilst here, improve the error message for when we do fail the dir
check.

Fixes: #776

Signed-off-by: Graham Whaley <graham.whaley@intel.com>