metrics: report: update to parse nested JSON
Now we nest the test results into uniquely named sub-structs
of the JSON, we need to un-nest it in the report processing in
order to get at the data.

We specifically un-nest, rather than just educate the scripts about
the sub-section names, as then the inner processing loops can stay
fairly generic across all the test results. It just makes the inner
code a bit less complex.

Signed-off-by: Graham Whaley <graham.whaley@intel.com>