metrics: json: enable socat and single line transmit for json
For some storage endpoints (such as the filebeat TCP), it is
much easier to use `socat` to send the data, rather than trying
to figure out how to get `curl` to do it.

Also, ideally we should send the results as a single line, so
add that capability as well.

And finally, document the hooks, and add a ToC to the README.

Fixes: #790

Signed-off-by: Graham Whaley <graham.whaley@intel.com>